backend/requirements.txt
fastapi
uvicorn
supabase>=2.0.0
pydantic-settings
sqlalchemy
psycopg2-binary
